Added an extra layer to the dashboard dispatch so it can accept a list of actions to dispatch. Separated the actions function collection list to it's own file so it could remain in memory. state and payload now must be accepted as arguments for each method. Progress commit on current work to reconnect student feed messaging.
