+++ client/src/pages/Dashboard/components/StudentCard/index.js
@@ -10,7 +10,7 @@ import api from "utils/api";
 import Icon from "components/Icon";
 import Dropdown from "components/Dropdown";
 import { getDashboardAction as gda, useClassroom, useDashboardDispatch, useStaffMember } from "pages/Dashboard/store";
-import { EDIT_STUDENT, REMOVE_STUDENT } from "pages/Dashboard/store/actions";
+import { EDIT_STUDENT, REMOVE_STUDENT } from "pages/Dashboard/store/actionsNames";
 import { usePriorityLevel } from "pages/Dashboard/utils/student";
 
 import "./style.sass";
