+++ client/src/pages/Dashboard/components/StudentModal/components/ActivityFeed/index.js
@@ -1,7 +1,5 @@
 import { useStudentFeed, useStudentFeedLoader } from "pages/Dashboard/store";
-import React/*, { useEffect, useReducer }*/ from "react";
-// import api from "utils/api";
-// import { useSocket } from "utils/socket.io";
+import React from "react";
 import CommentForm from "./components/CommentForm";
 
 import {
@@ -27,8 +25,6 @@ const ActivtyFeed = ({ student }) => {
     useStudentFeedLoader(feed);
     const entries = useStudentFeed();
 
-    // const socket = useSocket();
-
     const feedEntryComponentMap = item => {
 
         if( !typeMap[item.action] ) return null;
@@ -38,33 +34,6 @@ const ActivtyFeed = ({ student }) => {
         return <Entry key={item._id} student={student} by={item.by} date={item.date} data={item.data} />
 
     }
-
-    // useEffect(() => {
-
-    //     const handleFeedPush = message => dispatch(getAction("add",message.payload));
-
-    //     // socket.on( `rockfish:fish`, message => console.log("inner",message) );
-    //     // socket.on( `feedpush:${feed}`, handleFeedPush );
-
-    //     return () => socket.off( `feedpush:${feed}`, handleFeedPush );
-
-    // }, [feed, socket, dispatch]);
-
-    // useEffect(() => {
-
-    //     const getItems = async () => dispatch(getAction( "set", (await api.getFeedItems(feed)).data ));
-
-    //     try {
-
-    //         getItems();
-
-    //     } catch(err) {
-
-    //         console.log( err );
-
-    //     }
-
-    // },[feed]);
 
     return (
         entries &&
