+++ client/src/pages/Dashboard/components/Topbar/index.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import "./style.sass";
 import { useDashboardContext, getDashboardAction as gda } from "pages/Dashboard/store";
-import { SET_TOPBAR } from "pages/Dashboard/store/actions";
+import { SET_TOPBAR } from "pages/Dashboard/store/actionsNames";
 
 import { Heading } from "react-bulma-components";
