+++ client/src/pages/Dashboard/store/actionsNames.js
@@ -1,0 +1,44 @@
+/**
+ * TOPBAR ACTIONS
+ */
+export const SET_TOPBAR = "SET_TOPBAR";
+
+/**
+ * CLASSROOM ACTIONS
+ */
+export const SET_CLASSROOM = "SET_CLASSROOM";
+
+/**
+ * STAFF ACTIONS
+ */
+export const ADD_STAFF = "ADD_STAFF";
+export const REMOVE_STAFF = "REMOVE_STAFF";
+
+/**
+ * INVITE ACTIONS
+ */
+export const ADD_INVITE = "ADD_INVITE";
+export const DELETE_INVITE = "DELETE_INVITE";
+
+/**
+ * APP ACTIONS
+ */
+export const SET_MANAGE_APPS = "SET_MANAGE_APPS";
+export const ADD_APP = "ADD_APP";
+
+/**
+ * STUDENT VIEW ACTIONS
+ */
+export const SET_STUDENTS = "SET_STUDENTS";
+export const ADD_STUDENT = "ADD_STUDENT";
+export const ADD_STUDENTS = "ADD_STUDENTS";
+export const UPDATE_STUDENT = "UPDATE_STUDENT";
+export const REMOVE_STUDENT = "REMOVE_STUDENT";
+
+/**
+ * STUDENT MODAL ACTIONS
+ */
+export const EDIT_STUDENT = "EDIT_STUDENT";
+export const SET_STUDENT_FEED = "SET_STUDENT_FEED";
+export const ADD_STUDENT_FEED_ITEM = "ADD_STUDENT_FEED_ITEM";
+export const ADD_STUDENT_FEED_ITEMS = "ADD_STUDENT_FEED_ITEMS";
