+++ client/src/pages/Dashboard/utils/apps.js
@@ -1,5 +1,5 @@
 import { useDashboardDispatch, getDashboardAction as gda } from "pages/Dashboard/store"
-import { SET_MANAGE_APPS } from "pages/Dashboard/store/actions";
+import { SET_MANAGE_APPS } from "pages/Dashboard/store/actionsNames";
 
 export const useManageApps = () => {
