+++ client/src/pages/Dashboard/views/Team/components/InviteModal.js
@@ -11,7 +11,7 @@ import Icon from "components/Icon";
 import Form, { createValidator } from "components/Form";
 import api from "utils/api";
 import { useClassroom, useDashboardDispatch, getDashboardAction as gda } from "pages/Dashboard/store";
-import { ADD_INVITE } from "pages/Dashboard/store/actions";
+import { ADD_INVITE } from "pages/Dashboard/store/actionsNames";
 import { useSocket } from "utils/socket.io";
 
 const validateInviteData = createValidator({
